divided the old js files into better ones, using import/export
