+++ index.html
@@ -25,10 +25,11 @@
     <link rel="stylesheet" href="css/general.css" />
     <link rel="stylesheet" href="css/footer.css" />
 
-    <script src="javascript/generalrecipe.js" defer></script>
-    <script src="javascript/index.js" defer></script>
-    <script src="javascript/searchfunction+loadbtn.js" defer></script>
-    <script src="javascript/heartclick.js" defer></script>
+    <script type="module" src="javascript/dropdown.js" defer></script>
+    <script type="module" src="javascript/favorites.js" defer></script>
+    <script type="module" src="javascript/popularRecipes.js" defer></script>
+    <script type="module" src="javascript/recipeService.js" defer></script>
+    <script type="module" src="javascript/searchAndFilter.js" defer></script>
   </head>
   <body>
     <!--Navigation-->
