--- javascript/allmealspage.js
@@ -1,1 +1,0 @@
-// allmealspgae script
