--- javascript/chosenrecipe.js
@@ -1,1 +1,0 @@
-// chosenrecipe script
