--- javascript/dinnerpage.js
@@ -1,1 +1,0 @@
-// dinnerpage script
