+++ javascript/dropdown.js
@@ -1,0 +1,15 @@
+document.addEventListener("DOMContentLoaded", () => {
+    const dropdownBtn = document.getElementById("mealsDropdownBtn");
+    const dropdownMenu = document.getElementById("mealsDropdownMenu");
+  
+    dropdownBtn.addEventListener("click", () => {
+      dropdownMenu.classList.toggle("hidden");
+    });
+  
+    //Hides the menu if clicking outside
+    document.addEventListener("click", (event) => {
+      if (!dropdownBtn.contains(event.target) && !dropdownMenu.contains(event.target)) {
+        dropdownMenu.classList.add("hidden");
+      }
+    });
+  });
